requirements.txt
albumentations==0.4.6
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2019.11.28
chardet==4.0.0
Click==7.0
colorama==0.4.4
comtypes==1.1.7
cycler==0.10.0
DateTime==4.3
distlib==0.3.1
efficientnet-pytorch==0.7.1
ExifRead==2.3.2
filelock==3.8.0
Flask==1.1.1
Flask-Cors==3.0.10
Flask-WTF==0.14.3
geographiclib==1.50
geopy==2.1.0
gpsphoto==2.2.3
gunicorn==19.6.0
html5lib==1.1
idna==2.10
imageio==2.16.1
imgaug==0.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.13.2
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.4.2
networkx==2.7.1
nltk==3.7
numpy==1.20.1
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.2.4
piexif==1.1.3
Pillow==9.1.0
protobuf==3.15.8
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.0
pyttsx3==2.90
pytz==2019.3
PyWavelets==1.3.0
pywin32==228
PyYAML==6.0
regex==2022.9.13
requests==2.25.1
scikit-image==0.19.2
scikit-learn==1.1.2
scipy==1.6.3
Shapely==1.8.1.post1
six==1.14.0
sklearn==0.0
soupsieve==2.2.1
textblob==0.17.1
threadpoolctl==3.1.0
tifffile==2022.3.25
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.1
typing-extensions==3.7.4.3
urllib3==1.26.4
vboxapi==1.0
virtualenv==20.3.0
webencodings==0.5.1
Werkzeug==1.0.0
WTForms==2.3.1
zope.interface==5.4.0
